Add microdata for CPS
