+++ openfisca_us/data/__init__.py
@@ -1,0 +1,1 @@
+from openfisca_us.data.datasets import CPS, RawCPS
