+++ openfisca_us/data/datasets/__init__.py
@@ -1,0 +1,3 @@
+from openfisca_us.data.datasets.cps import CPS, RawCPS
+
+DATASETS = [CPS, RawCPS]
