+++ openfisca_us/data/storage/README.md
@@ -1,0 +1,3 @@
+# OpenFisca-US microdata
+
+This folder will contain microdata files (.h5) to support OpenFisca-US's microsimulation runs.
