+++ openfisca_us/data/storage/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+OPENFISCA_US_MICRODATA_FOLDER = Path(__file__).parent
