--- openfisca_us/microdata/__init__.py
@@ -1,6 +1,0 @@
-from openfisca_us.microdata.datasets import *
-from pathlib import Path
-
-REPO = Path(__file__).parent
-
-DATASETS = (RawCPS, CPS, RawACS, ACS, RawCE, CE)
