--- openfisca_us/microdata/datasets/__init__.py
@@ -1,3 +1,0 @@
-from openfisca_us.microdata.datasets.cps import *
-from openfisca_us.microdata.datasets.acs import *
-from openfisca_us.microdata.datasets.ce import *
