--- openfisca_us/microdata/datasets/acs/__init__.py
@@ -1,2 +1,0 @@
-from openfisca_us.microdata.datasets.acs.raw_acs import RawACS
-from openfisca_us.microdata.datasets.acs.acs import ACS
