--- openfisca_us/microdata/datasets/ce/__init__.py
@@ -1,2 +1,0 @@
-from openfisca_us.microdata.datasets.ce.raw_ce import RawCE
-from openfisca_us.microdata.datasets.ce.ce import CE
