--- openfisca_us/microdata/datasets/ce/emission_coefficients.yaml
@@ -1,33 +1,0 @@
-# The Impact of a Carbon Tax on Inequality - Supplemental Materials
-# Table A1: Carbon Intensity of Industries (in kgCO2/$)
-# Anders Fremstad & Mark Paul (2019)
-# https://docs.google.com/document/d/1jQlG1uz1UAPy6O_fmxyLQxRKLRDRpp5qmYdXKPXXXmM/edit
-# Accessed 2021-11-30
-
-airfare: 1.0
-alcohol: 0.33
-education: 0.24
-auto_insurance: 0.05
-autos: 0.73
-books: 0.22
-charity: 0.19
-clothes: 0.22
-electricity: 2.24
-food_at_home: 0.39
-food_at_restaurants: 0.24
-furnishings: 0.71
-gasoline: 3.22
-health: 0.22
-home_heating_fuel: 2.75
-household_supplies: 0.36
-life_insurance: 0.05
-mass_transit: 0.94
-natural_gas: 1.82
-other_car_services: 0.23
-other_dwelling_rentals: 0.06
-other_recreation: 0.25
-recreation_and_sports: 0.70
-telephone: 0.18
-tentant_occupied_dwellings: 0.05
-tobacco: 0.36
-water: 0.38
