--- openfisca_us/microdata/datasets/cps/__init__.py
@@ -1,2 +1,0 @@
-from openfisca_us.microdata.datasets.cps.raw_cps import RawCPS
-from openfisca_us.microdata.datasets.cps.cps import CPS
