--- openfisca_us/microdata/microdata/README.md
@@ -1,3 +1,0 @@
-# microdata
-
-This folder will contain `.h5` files that can be used as inputs to OpenFisca-US.
