--- openfisca_us/microdata/microdata/openfisca_us/README.md
@@ -1,3 +1,0 @@
-# openfisca-us
-
-This folder will contain microdata (`*.h5`) files that can be used as input to OpenFisca-US.
