--- openfisca_us/microdata/tests/cps/taxcalc_cps.yml
@@ -1,136 +1,0 @@
-DSI:
-  2020: 189927269.0
-EIC:
-  2020: 9228312693.0
-FLPDYR:
-  2020: 41355857382607.0
-MARS:
-  2020: 34571710317.0
-RECID:
-  2020: 2868868782968629.0
-XTOT:
-  2020: 38912156785.0
-a_lineno:
-  2020: 32607341954.0
-age_head:
-  2020: 961076447150.0
-age_spouse:
-  2020: 296862623734.0
-agi_bin:
-  2020: 122252399747.0
-blind_head:
-  2020: 455822086.0
-blind_spouse:
-  2020: 79696777.0
-e00200:
-  2020: 728084861887664.0
-e00200p:
-  2020: 533442392509206.0
-e00200s:
-  2020: 194642469378458.0
-e00300:
-  2020: 9022412063844.0
-e00400:
-  2020: 6799848823821.0
-e00600:
-  2020: 33665562050123.0
-e00650:
-  2020: 17673043155149.0
-e00800:
-  2020: 745512001447.0
-e00900:
-  2020: 40874145404810.0
-e00900p:
-  2020: 29300533011638.0
-e00900s:
-  2020: 11573612393172.0
-e01100:
-  2020: 37621808271841.0
-e01400:
-  2020: 40537762955871.0
-e01500:
-  2020: 71170487877568.0
-e01700:
-  2020: 56626394906090.0
-e02100:
-  2020: 2806091696353.0
-e02100p:
-  2020: 2161542596415.0
-e02100s:
-  2020: 644549099938.0
-e02300:
-  2020: 3600730286426.0
-e02400:
-  2020: 118761526836328.0
-e03150:
-  2020: 108919590503.0
-e03210:
-  2020: 1082685811013.0
-e03240:
-  2020: 1615276480218.0
-e03270:
-  2020: 3264573234362.0
-e03300:
-  2020: 4391018461360.0
-e17500:
-  2020: 21884636513233.0
-e18400:
-  2020: 26905528743895.0
-e18500:
-  2020: 23405200664512.0
-e19200:
-  2020: 64135875257805.0
-e19800:
-  2020: 24847395932735.0
-e20100:
-  2020: 5447401989650.0
-e20400:
-  2020: 12632324212962.0
-e32800:
-  2020: 409107490379.0
-elderly_dependents:
-  2020: 2270452.0
-f2441:
-  2020: 6972399400.0
-ffpos:
-  2020: 23655117749.0
-fips:
-  2020: 566960272537.0
-h_seq:
-  2020: 1021879890905688.0
-housing_ben:
-  2020: 4745119846361.0
-mcaid_ben:
-  2020: 45414897893509.0
-mcare_ben:
-  2020: 69300599012862.0
-n1820:
-  2020: 1451955002.0
-n21:
-  2020: 25211397507.0
-n24:
-  2020: 9161751099.0
-nu06:
-  2020: 2988216945.0
-nu13:
-  2020: 6972399400.0
-nu18:
-  2020: 9825597437.0
-other_ben:
-  2020: 30496389219821.0
-pencon_p:
-  2020: 0.0
-pencon_s:
-  2020: 0.0
-s006:
-  2020: 1827818011056800.0
-snap_ben:
-  2020: 12859336560730.0
-ssi_ben:
-  2020: 6491584460372.0
-tanf_ben:
-  2020: 3654438465163.0
-vet_ben:
-  2020: 16466524730428.0
-wic_ben:
-  2020: 474671919174.0
