--- openfisca_us/microdata/tests/test_imports.py
@@ -1,26 +1,0 @@
-def test_import():
-    import openfisca_us.microdata
-
-
-def test_RawCPS_import():
-    from openfisca_us.microdata import RawCPS
-
-
-def test_CPS_import():
-    from openfisca_us.microdata import CPS
-
-
-def test_RawACS_import():
-    from openfisca_us.microdata import RawACS
-
-
-def test_ACS_import():
-    from openfisca_us.microdata import ACS
-
-
-def test_RawCE_import():
-    from openfisca_us.microdata import RawCE
-
-
-def test_CE_import():
-    from openfisca_us.microdata import CE
