+++ openfisca_us/tests/microsimulation/data/test_imports.py
@@ -1,0 +1,6 @@
+def test_RawCPS_import():
+    from openfisca_us.data import RawCPS
+
+
+def test_CPS_import():
+    from openfisca_us.data import CPS
